+++ lib/stealth/controller/dev_jumps.rb
@@ -5,14 +5,16 @@ module Stealth
   class Controller
     module DevJumps
 
+      DEV_JUMP_REGEX = /\A\/(.*)\/(.*)\z|\A\/\/(.*)\z|\A\/(.*)\z/
+
       extend ActiveSupport::Concern
 
       included do
         private
 
         def dev_jump_detected?
-          if Stealth.env == 'development'
-            if current_message.message&.match(/\/(.*)\/(.*)|\/\/(.*)|\/(.*)/)
+          if Stealth.env.development?
+            if current_message.message&.match(DEV_JUMP_REGEX)
               handle_dev_jump
               return true
             end
@@ -24,6 +26,11 @@ module Stealth
         def handle_dev_jump
           _, flow, state = current_message.message.split('/')
           flow = nil if flow.blank?
+
+          Stealth::Logger.l(
+            topic: 'dev_jump',
+            message: "Dev Jump detected: Flow: #{flow.inspect} State: #{state.inspect}"
+          )
 
           step_to flow: flow, state: state
         end