+++ lib/stealth/generators/builder/bot/controllers/unrecognized_messages_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class UnrecognizedMessagesController < BotController
+
+  def handle_unrecognized_message
+    do_nothing
+  end
+
+end