+++ lib/stealth/generators/builder/config/flow_map.rb
@@ -10,6 +10,14 @@ class FlowMap
     state :say_goodbye
   end
 
+  flow :interrupt do
+    state :say_interrupted
+  end
+
+  flow :unrecognized_message do
+    state :handle_unrecognized_message
+  end
+
   flow :catch_all do
     state :level1
   end