+++ lib/stealth/generators/builder/config/puma.rb
@@ -5,6 +5,21 @@ threads threads_count, threads_count
 #
 port ENV.fetch("PORT") { 3000 }
 
+# Specifies the number of `workers` to boot in clustered mode.
+# Workers are forked webserver processes. If using threads and workers together
+# the concurrency of the application would be max `threads` * `workers`.
+# Workers do not work on JRuby or Windows (both of which do not support
+# processes).
+#
+# workers ENV.fetch("WEB_CONCURRENCY") { 2 }
+
+# Use the `preload_app!` method when specifying a `workers` number.
+# This directive tells Puma to first boot the application and load code
+# before forking the application. This takes advantage of Copy On Write
+# process behavior so workers use less memory.
+#
+# preload_app!
+
 # Specifies the `environment` that Puma will run in.
 #
 environment ENV.fetch("STEALTH_ENV") { "development" }