+++ spec/controller/catch_all_spec.rb
@@ -1,26 +1,13 @@
-# coding: utf-8
 # frozen_string_literal: true
 
-require File.expand_path(File.join(File.dirname(__FILE__), '..', '/spec_helper'))
+require 'spec_helper'
 
 describe "Stealth::Controller::CatchAll" do
-
-  class VadersController < Stealth::Controller
-    def my_action
-      raise "oops"
-    end
-
-    def my_action2
-
-    end
-
-    def my_action3
-      do_nothing
-    end
-  end
+  $msg = nil
 
   class StubbedCatchAllsController < Stealth::Controller
     def level1
+      $msg = current_message
       do_nothing
     end
 
@@ -40,6 +27,8 @@ describe "Stealth::Controller::CatchAll" do
       state :my_action
       state :my_action2
       state :my_action3
+      state :action_with_unrecognized_msg
+      state :action_with_unrecognized_match
     end
 
     flow :catch_all do
@@ -101,39 +90,64 @@ describe "Stealth::Controller::CatchAll" do
     end
 
     describe "catch_alls from within catch_all flow" do
+      let(:e) {
+        e = OpenStruct.new
+        e.class = RuntimeError
+        e.message = 'oops'
+        e.backtrace = [
+          '/stealth/lib/stealth/controller/controller.rb',
+          '/stealth/lib/stealth/controller/catch_all.rb',
+        ]
+        e
+      }
+
       before(:each) do
         controller.current_session.session = Stealth::Session.canonical_session_slug(flow: 'catch_all', state: 'level1')
       end
 
       it "should not step_to to catch_all" do
         expect(controller).to_not receive(:step_to)
-        controller.run_catch_all
+        controller.run_catch_all(err: e)
       end
 
       it "should return false" do
-        expect(controller.run_catch_all).to be false
+        expect(controller.run_catch_all(err: e)).to be false
       end
 
       it "should log the error message" do
-        expect(Stealth::Logger).to receive(:l).with(topic: 'catch_all', message: "CatchAll level1 triggered for error-#{controller.current_session_id}-catch_all-level1: ")
+        expect(Stealth::Logger).to receive(:l).with(topic: 'catch_all', message: "[Level 1] oops\n/stealth/lib/stealth/controller/controller.rb\n/stealth/lib/stealth/controller/catch_all.rb")
         expect(Stealth::Logger).to receive(:l).with(topic: 'catch_all', message: 'CatchAll triggered from within CatchAll; ignoring.')
-        controller.run_catch_all
+        controller.run_catch_all(err: e)
       end
     end
 
-    describe "releasing locks" do
-      it "should release the session lock after the maximum number of catch_all levels have been reached" do
-        allow(controller).to receive(:fetch_error_level).and_return(4)
-        session = Stealth::Session.new(id: controller.current_session_id)
-        session.set_session(new_flow: 'vader', new_state: 'my_action')
-        expect(controller).to receive(:release_lock!)
-        controller.run_catch_all
+    describe "catch_all_reason" do
+      before(:each) do
+        @session = Stealth::Session.new(id: controller.current_session_id)
+        @session.set_session(new_flow: 'vader', new_state: 'my_action2')
       end
 
-      it "should release the session lock if the bot does not have a CatchAll flow" do
-        FlowMap.flow_spec[:catch_all] = nil
-        expect(controller).to receive(:release_lock!)
-        controller.run_catch_all
+      after(:each) do
+        $msg = nil
+      end
+
+      it 'should have access to the error raised in current_message.catch_all_reason' do
+        controller.action(action: :my_action)
+        expect($msg.catch_all_reason).to be_a(Hash)
+        expect($msg.catch_all_reason[:err]).to eq(RuntimeError)
+        expect($msg.catch_all_reason[:err_msg]).to eq('oops')
+      end
+
+      it 'should have the correct error when handle_message fails to recognize a message' do
+        controller.action(action: :action_with_unrecognized_msg)
+        expect($msg.catch_all_reason[:err]).to eq(Stealth::Errors::UnrecognizedMessage)
+        expect($msg.catch_all_reason[:err_msg]).to eq("The reply '#{facebook_message.message_with_text.message}' was not recognized.")
+      end
+
+      it 'should have the correct error when get_match fails to recognize a message' do
+        controller.action(action: :action_with_unrecognized_match)
+        expect($msg.catch_all_reason[:err]).to eq(Stealth::Errors::UnrecognizedMessage)
+        expect($msg.catch_all_reason[:err_msg]).to eq("The reply '#{facebook_message.message_with_text.message}' was not recognized.")
       end
     end
   end