+++ spec/scheduled_reply_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require File.expand_path(File.dirname(__FILE__) + '/spec_helper')
+require 'spec_helper'
 
 describe "Stealth::ScheduledReplyJob" do