+++ spec/support/nlp_clients/dialogflow.rb
@@ -1,0 +1,9 @@
+module Stealth
+  module Nlp
+    module Dialogflow
+      class Client < Stealth::Nlp::Client
+
+      end
+    end
+  end
+end